+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include README.rst
-include src/PvMail/LICENSE.txt
+include src/PvMail/LICENSE
 include src/PvMail/resources/*
