+++ src/PvMail/tests/FIXME-test_mailer.py
@@ -9,7 +9,7 @@ unit tests for the spec module
 #
 # Distributed under the terms of the Creative Commons Attribution 4.0 International Public License.
 #
-# The full license is in the file LICENSE.txt, distributed with this software.
+# The full license is in the file LICENSE, distributed with this software.
 # -----------------------------------------------------------------------------
